move costfns to common file, delete old costfns, show different costfn use in lineup play
